+++ src/RelationshipMatrices.jl
@@ -12,8 +12,9 @@ include("kinship.jl")
 include("ainv.jl")
 include("grm.jl")
 include("grm-bits.jl")
+include("hinv.jl")
 include("irm.jl")
 
-export nrm, nrm_diag, Ainv, ainv, grm, kinship, validate_pedigree
+export nrm, nrm_diag, Ainv, ainv, grm, kinship, validate_pedigree, hinv, Hinv
 
 end # module RelationshipMatrices